pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "docbrain"
version = "0.3.0"
description = "Azure-native AI Knowledge Base with RAG Search — Built on Azure AI Foundry, Azure AI Search, and Azure Blob Storage"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Praveen", email = "rpraveen760@gmail.com"},
]

dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "jinja2>=3.1.0",

    # Azure AI Foundry SDK
    "azure-ai-projects>=2.0.0",
    "azure-identity>=1.16.0",
    "openai>=1.12.0",

    # Azure AI Search
    "azure-search-documents>=11.6.0",

    # Azure Blob Storage
    "azure-storage-blob>=12.19.0",

    # Document processing
    "tiktoken>=0.6.0",
    "pymupdf>=1.24.0",
    "markdown>=3.5.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
